Move ColumnEvaluations to berkeley_columns
