+++ kimchi/src/circuits/berkeley_columns.rs
@@ -1,7 +1,10 @@
-use crate::circuits::{
-    expr::{self, Domain, GenericColumn},
-    gate::{CurrOrNext, GateType},
-    lookup::lookups::LookupPattern,
+use crate::{
+    circuits::{
+        expr::{self, ColumnEvaluations, Domain, ExprError, GenericColumn},
+        gate::{CurrOrNext, GateType},
+        lookup::lookups::LookupPattern,
+    },
+    proof::{PointEvaluations, ProofEvaluations},
 };
 use serde::{Deserialize, Serialize};
 use CurrOrNext::{Curr, Next};
@@ -88,6 +91,69 @@ impl expr::Variable<Column> {
         match self.row {
             Curr => format!("Curr({col})"),
             Next => format!("Next({col})"),
+        }
+    }
+}
+
+impl<F: Copy> ColumnEvaluations<F> for ProofEvaluations<PointEvaluations<F>> {
+    type Column = Column;
+    fn evaluate(&self, col: Self::Column) -> Result<PointEvaluations<F>, ExprError<Self::Column>> {
+        use Column::*;
+        match col {
+            Witness(i) => Ok(self.w[i]),
+            Z => Ok(self.z),
+            LookupSorted(i) => self.lookup_sorted[i].ok_or(ExprError::MissingIndexEvaluation(col)),
+            LookupAggreg => self
+                .lookup_aggregation
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            LookupTable => self
+                .lookup_table
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            LookupRuntimeTable => self
+                .runtime_lookup_table
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            Index(GateType::Poseidon) => Ok(self.poseidon_selector),
+            Index(GateType::Generic) => Ok(self.generic_selector),
+            Index(GateType::CompleteAdd) => Ok(self.complete_add_selector),
+            Index(GateType::VarBaseMul) => Ok(self.mul_selector),
+            Index(GateType::EndoMul) => Ok(self.emul_selector),
+            Index(GateType::EndoMulScalar) => Ok(self.endomul_scalar_selector),
+            Index(GateType::RangeCheck0) => self
+                .range_check0_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            Index(GateType::RangeCheck1) => self
+                .range_check1_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            Index(GateType::ForeignFieldAdd) => self
+                .foreign_field_add_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            Index(GateType::ForeignFieldMul) => self
+                .foreign_field_mul_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            Index(GateType::Xor16) => self
+                .xor_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            Index(GateType::Rot64) => self
+                .rot_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            Permutation(i) => Ok(self.s[i]),
+            Coefficient(i) => Ok(self.coefficients[i]),
+            LookupKindIndex(LookupPattern::Xor) => self
+                .xor_lookup_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            LookupKindIndex(LookupPattern::Lookup) => self
+                .lookup_gate_lookup_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            LookupKindIndex(LookupPattern::RangeCheck) => self
+                .range_check_lookup_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            LookupKindIndex(LookupPattern::ForeignFieldMul) => self
+                .foreign_field_mul_lookup_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            LookupRuntimeSelector => self
+                .runtime_lookup_table_selector
+                .ok_or(ExprError::MissingIndexEvaluation(col)),
+            Index(_) => Err(ExprError::MissingIndexEvaluation(col)),
         }
     }
 }
